pyproject.toml
[project]
name = "sentiment-agent"
version = "0.1.0"
description = "Sentiment analysis agent using ADK, MCP and A2A"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk==0.2.8",
    "fastmcp==2.9.2",
    "google-adk==1.5.0",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
]
